requirements.txt
Django              3.2.5
django-crispy-forms 1.12.0
django-redis        5.0.0
web3                5.21.0

